backend/requirements.txt
fastapi
uvicorn[standard]
pymupdf
google-generativeai
python-multipart
python-dotenv
